+++ app/Http/Controllers/Modrinth/TeamController.php
@@ -11,46 +11,44 @@ class TeamController extends Controller
     const TEAM_ID = 'membercat';
     const CACHE_KEYS = [
         'team' => 'modrinth:team:' . self::TEAM_ID,
-        'members' => 'modrinth:team:' . self::TEAM_ID . ':members',
     ];
+
+    private function fetchFromApi($endpoint, $errorMessage)
+    {
+        $response = Http::get("https://api.modrinth.com/v3/organization/{$endpoint}");
+        
+        if ($response->successful()) {
+            return $response->json();
+        }
+
+        return response()->json(['error' => $errorMessage], 500);
+    }
 
     public function fetchRawData()
     {
         $cacheKey = self::CACHE_KEYS['team'];
         $response = Cache::remember($cacheKey, 300, function () { // 5 mins
-            $response = Http::get('https://api.modrinth.com/v3/organization/membercat');
-
-            if ($response->successful()) {
-                return $response->json();
-            }
-
-            return null;
+            return $this->fetchFromApi('membercat', 'Unable to fetch team data');
         });
 
-        if ($response) {
+        if (!isset($response['error'])) {
             return $response;
         }
 
-        return response()->json(['error' => 'Unable to fetch team data'], 500);
+        return $response;
     }
 
     public function fetchMembers()
     {
-        $cacheKey = self::CACHE_KEYS['members'];
-        $response = Cache::remember($cacheKey, 600, function () { // 10 mins
-            $response = Http::get("https://api.modrinth.com/v3/organization/membercat/members");
-
-            if ($response->successful()) {
-                return $response->json();
-            }
-
-            return null;
+        $cacheKey = self::CACHE_KEYS['team'] . ':members';
+        $response = Cache::remember($cacheKey, 600, function () {
+            return $this->fetchFromApi('membercat/members', 'Unable to fetch members');
         });
 
-        if ($response) {
+        if (!isset($response['error'])) {
             return $response;
         }
 
-        return response()->json(['error' => 'Unable to fetch members'], 500);
+        return $response;
     }
 }